Do not instrument bowerick.java-interfaces for cloverage.
The problem observed in commit 2650a00ad97fc0415512c251b602c3dbf8f9f318 was:
      actual: java.lang.LinkageError: loader constraint violation: when resolving method 'java.lang.AutoCloseable bowerick.JmsController.createJsonConsumer(java.lang.String, java.lang.String, bowerick.JmsConsumerCallback, int)' the class loader clojure.lang.DynamicClassLoader @3df99d5e of the current class, bowerick/test/java_interop$fn__5146, and the class loader 'app' for the method's defining class, bowerick/JmsController, have different Class objects for the type bowerick/JmsConsumerCallback used in the signature (bowerick.test.java_interop$fn__5146 is in unnamed module of loader clojure.lang.DynamicClassLoader @3df99d5e, parent loader clojure.lang.DynamicClassLoader @5d95c7c2; bowerick.JmsController is in unnamed module of loader 'app')
     at bowerick.test.java_interop$fn__5146.invokeStatic (java_interop.clj:96)

It looks like that cloverage uses a different class loader for the
intrumented code. This seems to have caused a problem when implementing
the JmsConsumerCallback interface in the test function, which seems to
have used a different class loader than the class loader that was used
for the instrumented code.
Avoiding intstrumentation of bowerick.java-interfaces fixes the issue
with cloverage.